Split practice-page-logic.test.ts by user flow
Delete the source-size guard because Track B replaces it with repo-wide test file size warnings.
