--- app/(app)/app/practice/practice-page-logic.test.ts
@@ -1,1756 +1,0 @@
-// @vitest-environment jsdom
-import { readFileSync } from 'node:fs';
-import path from 'node:path';
-import { describe, expect, it, vi } from 'vitest';
-import {
-  canSubmitAnswer,
-  createBookmarksEffect,
-  createDifficultyChangeHandler,
-  createLoadNextQuestionAction,
-  createSessionCountBlurHandler,
-  createSessionCountChangeHandler,
-  createSessionModeChangeHandler,
-  createStatusChangeHandler,
-  createToggleTagHandler,
-  handleSessionCountBlur,
-  handleSessionCountChange,
-  handleSessionModeChange,
-  loadNextQuestion,
-  SESSION_COUNT_MAX,
-  SESSION_COUNT_MIN,
-  selectChoiceIfAllowed,
-  startSession,
-  submitAnswerForQuestion,
-  toggleBookmarkForQuestion,
-} from '@/app/(app)/app/practice/practice-page-logic';
-import { toPracticeSessionRoute } from '@/lib/routes';
-import type { ActionResult } from '@/src/adapters/controllers/action-result';
-import { err, ok } from '@/src/adapters/controllers/action-result';
-import { createNextQuestion } from '@/src/application/test-helpers/create-next-question';
-import type { NextQuestion } from '@/src/application/use-cases/get-next-question';
-import type { SubmitAnswerOutput } from '@/src/application/use-cases/submit-answer';
-import { createDeferred } from '@/tests/test-helpers/create-deferred';
-
-describe('practice-page-logic', () => {
-  it('keeps practice-page-logic.ts within the SPEC-020 line cap', () => {
-    const filePath = path.resolve(
-      process.cwd(),
-      'app/(app)/app/practice/practice-page-logic.ts',
-    );
-    const lineCount = readFileSync(filePath, 'utf8')
-      .trimEnd()
-      .split('\n').length;
-
-    expect(lineCount).toBeLessThanOrEqual(300);
-  });
-
-  describe('canSubmitAnswer', () => {
-    it('returns true when question is loaded, choice is selected, and not loading', () => {
-      expect(
-        canSubmitAnswer({
-          loadState: { status: 'ready' },
-          question: createNextQuestion(),
-          selectedChoiceId: 'choice_1',
-          isAnswered: false,
-          submitResult: null,
-        }),
-      ).toBe(true);
-    });
-
-    it('returns false when loadState is loading', () => {
-      expect(
-        canSubmitAnswer({
-          loadState: { status: 'loading' },
-          question: createNextQuestion(),
-          selectedChoiceId: 'choice_1',
-          isAnswered: false,
-          submitResult: null,
-        }),
-      ).toBe(false);
-    });
-
-    it('returns false when submitResult exists', () => {
-      expect(
-        canSubmitAnswer({
-          loadState: { status: 'ready' },
-          question: createNextQuestion(),
-          selectedChoiceId: 'choice_1',
-          isAnswered: false,
-          submitResult: {
-            attemptId: 'attempt_1',
-            isCorrect: true,
-            correctChoiceId: 'choice_1',
-            explanationMd: 'Because…',
-            referenceMd: null,
-            choiceExplanations: [],
-          },
-        }),
-      ).toBe(false);
-    });
-
-    it('returns false when question is already answered', () => {
-      expect(
-        canSubmitAnswer({
-          loadState: { status: 'ready' },
-          question: createNextQuestion(),
-          selectedChoiceId: 'choice_1',
-          isAnswered: true,
-          submitResult: null,
-        }),
-      ).toBe(false);
-    });
-  });
-
-  describe('loadNextQuestion', () => {
-    it('ignores stale responses when a newer request finishes first', async () => {
-      const first = createDeferred<ActionResult<NextQuestion | null>>();
-      const second = createDeferred<ActionResult<NextQuestion | null>>();
-      let latestRequestId = 0;
-      const responseQueue = [first.promise, second.promise];
-
-      const getNextQuestionFn = vi.fn(async () => {
-        const nextResponse = responseQueue.shift();
-        if (!nextResponse) {
-          throw new Error('Unexpected call to getNextQuestionFn');
-        }
-        return nextResponse;
-      });
-
-      const setQuestion = vi.fn();
-      const setLoadState = vi.fn();
-
-      const createRequestSequenceId = () => {
-        latestRequestId += 1;
-        return latestRequestId;
-      };
-
-      const isLatestRequest = (requestId: number) =>
-        requestId === latestRequestId;
-
-      const loadFirst = loadNextQuestion({
-        getNextQuestionFn,
-        filters: { tagSlugs: [], difficulty: null, status: 'unanswered' },
-        createIdempotencyKey: () => 'idem_1',
-        nowMs: () => 1234,
-        setLoadState,
-        setSelectedChoiceId: vi.fn(),
-        setSubmitResult: vi.fn(),
-        setSubmitIdempotencyKey: vi.fn(),
-        setQuestionLoadedAt: vi.fn(),
-        setQuestion,
-        createRequestSequenceId,
-        isLatestRequest,
-      });
-
-      const loadSecond = loadNextQuestion({
-        getNextQuestionFn,
-        filters: { tagSlugs: [], difficulty: null, status: 'unanswered' },
-        createIdempotencyKey: () => 'idem_2',
-        nowMs: () => 5678,
-        setLoadState,
-        setSelectedChoiceId: vi.fn(),
-        setSubmitResult: vi.fn(),
-        setSubmitIdempotencyKey: vi.fn(),
-        setQuestionLoadedAt: vi.fn(),
-        setQuestion,
-        createRequestSequenceId,
-        isLatestRequest,
-      });
-
-      second.resolve(
-        ok(
-          createNextQuestion({
-            questionId: 'q_2',
-            slug: 'q-2',
-          }),
-        ),
-      );
-      await loadSecond;
-
-      first.resolve(ok(createNextQuestion()));
-      await loadFirst;
-
-      expect(setQuestion).toHaveBeenCalledTimes(1);
-      expect(setQuestion).toHaveBeenCalledWith(
-        expect.objectContaining({ questionId: 'q_2' }),
-      );
-      expect(setLoadState.mock.calls.at(-1)?.[0]).toEqual({ status: 'ready' });
-    });
-
-    it('loads next question and updates loadedAt when a question exists', async () => {
-      const getNextQuestionFn = vi.fn(async () => ok(createNextQuestion()));
-      const setLoadState = vi.fn();
-      const setSelectedChoiceId = vi.fn();
-      const setSubmitResult = vi.fn();
-      const setSubmitIdempotencyKey = vi.fn();
-      const setQuestionLoadedAt = vi.fn();
-      const setQuestion = vi.fn();
-
-      await loadNextQuestion({
-        getNextQuestionFn,
-        filters: {
-          tagSlugs: ['opioids'],
-          difficulty: 'easy',
-          status: 'unanswered',
-        },
-        createIdempotencyKey: () => 'idem_1',
-        nowMs: () => 1234,
-        setLoadState,
-        setSelectedChoiceId,
-        setSubmitResult,
-        setSubmitIdempotencyKey,
-        setQuestionLoadedAt,
-        setQuestion,
-      });
-
-      expect(getNextQuestionFn).toHaveBeenCalledWith({
-        filters: {
-          tagSlugs: ['opioids'],
-          difficulties: ['easy'],
-          statuses: ['unanswered'],
-        },
-      });
-      expect(setLoadState).toHaveBeenCalledWith({ status: 'loading' });
-      expect(setSelectedChoiceId).toHaveBeenCalledWith(null);
-      expect(setSubmitResult).toHaveBeenCalledWith(null);
-      expect(setSubmitIdempotencyKey).toHaveBeenCalledWith(null);
-      expect(setQuestionLoadedAt).toHaveBeenCalledWith(null);
-
-      expect(setQuestion).toHaveBeenCalledWith(
-        expect.objectContaining({ questionId: 'q_1' }),
-      );
-      expect(setQuestionLoadedAt).toHaveBeenCalledWith(1234);
-      expect(setSubmitIdempotencyKey).toHaveBeenLastCalledWith('idem_1');
-      expect(setLoadState).toHaveBeenCalledWith({ status: 'ready' });
-    });
-
-    it('sets loadedAt to null when there is no next question', async () => {
-      const setQuestionLoadedAt = vi.fn();
-      const setSubmitIdempotencyKey = vi.fn();
-
-      await loadNextQuestion({
-        getNextQuestionFn: async () => ok(null),
-        filters: { tagSlugs: [], difficulty: null, status: 'unanswered' },
-        createIdempotencyKey: () => 'idem_1',
-        nowMs: () => 1234,
-        setLoadState: vi.fn(),
-        setSelectedChoiceId: vi.fn(),
-        setSubmitResult: vi.fn(),
-        setSubmitIdempotencyKey,
-        setQuestionLoadedAt,
-        setQuestion: vi.fn(),
-      });
-
-      expect(setQuestionLoadedAt).toHaveBeenCalledWith(null);
-      expect(setSubmitIdempotencyKey).toHaveBeenLastCalledWith(null);
-    });
-
-    it('sets error state when controller fails', async () => {
-      const setLoadState = vi.fn();
-      const setQuestion = vi.fn();
-      const setSubmitIdempotencyKey = vi.fn();
-
-      await loadNextQuestion({
-        getNextQuestionFn: async () =>
-          err('UNSUBSCRIBED', 'Subscription required'),
-        filters: { tagSlugs: [], difficulty: null, status: 'unanswered' },
-        createIdempotencyKey: () => 'idem_1',
-        nowMs: () => 1234,
-        setLoadState,
-        setSelectedChoiceId: vi.fn(),
-        setSubmitResult: vi.fn(),
-        setSubmitIdempotencyKey,
-        setQuestionLoadedAt: vi.fn(),
-        setQuestion,
-      });
-
-      expect(setQuestion).toHaveBeenCalledWith(null);
-      expect(setSubmitIdempotencyKey).toHaveBeenLastCalledWith(null);
-      expect(setLoadState).toHaveBeenCalledWith({
-        status: 'error',
-        message: 'Subscription required',
-      });
-    });
-
-    it('sets error state when controller throws', async () => {
-      const setLoadState = vi.fn();
-      const setQuestionLoadedAt = vi.fn();
-      const setSubmitIdempotencyKey = vi.fn();
-      const setQuestion = vi.fn();
-
-      await loadNextQuestion({
-        getNextQuestionFn: async () => {
-          throw new Error('Network down');
-        },
-        filters: { tagSlugs: [], difficulty: null, status: 'unanswered' },
-        createIdempotencyKey: () => 'idem_1',
-        nowMs: () => 1234,
-        setLoadState,
-        setSelectedChoiceId: vi.fn(),
-        setSubmitResult: vi.fn(),
-        setSubmitIdempotencyKey,
-        setQuestionLoadedAt,
-        setQuestion,
-      });
-
-      expect(setQuestion).toHaveBeenCalledWith(null);
-      expect(setQuestionLoadedAt).toHaveBeenLastCalledWith(null);
-      expect(setSubmitIdempotencyKey).toHaveBeenLastCalledWith(null);
-      expect(setLoadState).toHaveBeenCalledWith({
-        status: 'error',
-        message: 'Network down',
-      });
-    });
-
-    it('returns no state updates when unmounted during loadNextQuestion', async () => {
-      const deferred = createDeferred<ActionResult<NextQuestion | null>>();
-      let mounted = true;
-
-      const setLoadState = vi.fn();
-      const setQuestionLoadedAt = vi.fn();
-      const setSubmitIdempotencyKey = vi.fn();
-      const setQuestion = vi.fn();
-
-      const promise = loadNextQuestion({
-        getNextQuestionFn: async () => deferred.promise,
-        filters: { tagSlugs: [], difficulty: null, status: 'unanswered' },
-        createIdempotencyKey: () => 'idem_1',
-        nowMs: () => 1234,
-        setLoadState,
-        setSelectedChoiceId: vi.fn(),
-        setSubmitResult: vi.fn(),
-        setSubmitIdempotencyKey,
-        setQuestionLoadedAt,
-        setQuestion,
-        isMounted: () => mounted,
-      });
-
-      mounted = false;
-      deferred.resolve(ok(createNextQuestion()));
-      await promise;
-
-      expect(setQuestion).not.toHaveBeenCalled();
-      expect(setQuestionLoadedAt).not.toHaveBeenCalledWith(1234);
-      expect(setSubmitIdempotencyKey).not.toHaveBeenCalledWith('idem_1');
-      expect(setLoadState).not.toHaveBeenCalledWith({ status: 'ready' });
-    });
-  });
-
-  describe('createLoadNextQuestionAction', () => {
-    it('runs load inside startTransition', () => {
-      const startTransition = vi.fn((fn: () => void) => fn());
-      const setLoadState = vi.fn();
-
-      const action = createLoadNextQuestionAction({
-        startTransition,
-        getNextQuestionFn: async () => ok(createNextQuestion()),
-        filters: { tagSlugs: [], difficulty: null, status: 'unanswered' },
-        createIdempotencyKey: () => 'idem_1',
-        nowMs: () => 1234,
-        setLoadState,
-        setSelectedChoiceId: vi.fn(),
-        setSubmitResult: vi.fn(),
-        setSubmitIdempotencyKey: vi.fn(),
-        setQuestionLoadedAt: vi.fn(),
-        setQuestion: vi.fn(),
-      });
-
-      action();
-
-      expect(startTransition).toHaveBeenCalledTimes(1);
-      expect(setLoadState).toHaveBeenCalledWith({ status: 'loading' });
-    });
-  });
-
-  describe('createBookmarksEffect', () => {
-    it('loads bookmarks and updates state on success', async () => {
-      const setBookmarkedQuestionIds = vi.fn();
-      const setBookmarkStatus = vi.fn();
-
-      const cleanup = createBookmarksEffect({
-        bookmarkRetryCount: 0,
-        getBookmarksFn: async () =>
-          ok({ rows: [{ questionId: 'q_1' }, { questionId: 'q_2' }] }),
-        setBookmarkedQuestionIds,
-        setBookmarkStatus,
-        setBookmarkRetryCount: vi.fn(),
-        logError: vi.fn(),
-      });
-
-      await new Promise((resolve) => setTimeout(resolve, 0));
-
-      expect(setBookmarkedQuestionIds).toHaveBeenCalledWith(
-        new Set(['q_1', 'q_2']),
-      );
-      expect(setBookmarkStatus).toHaveBeenNthCalledWith(1, 'loading');
-      expect(setBookmarkStatus).toHaveBeenNthCalledWith(2, 'idle');
-
-      cleanup();
-    });
-
-    it('retries when bookmarks load fails and retryCount < 2', async () => {
-      vi.useFakeTimers();
-      try {
-        let retry = 0;
-        const setBookmarkRetryCount = vi.fn(
-          (next: number | ((prev: number) => number)) => {
-            retry = typeof next === 'function' ? next(retry) : next;
-          },
-        );
-
-        const setTimeoutFn = vi.fn((fn: () => void, ms: number) =>
-          setTimeout(fn, ms),
-        );
-        const clearTimeoutFn = vi.fn((id: ReturnType<typeof setTimeout>) =>
-          clearTimeout(id),
-        );
-
-        const cleanup = createBookmarksEffect({
-          bookmarkRetryCount: 0,
-          getBookmarksFn: async () => err('INTERNAL_ERROR', 'Boom'),
-          setBookmarkedQuestionIds: vi.fn(),
-          setBookmarkStatus: vi.fn(),
-          setBookmarkRetryCount,
-          setTimeoutFn,
-          clearTimeoutFn,
-          logError: vi.fn(),
-        });
-
-        await vi.advanceTimersByTimeAsync(0);
-
-        expect(setTimeoutFn).toHaveBeenCalledWith(expect.any(Function), 1000);
-
-        await vi.advanceTimersByTimeAsync(1000);
-        expect(retry).toBe(1);
-
-        cleanup();
-      } finally {
-        vi.useRealTimers();
-      }
-    });
-
-    it('does not retry when retryCount >= 2', async () => {
-      const setTimeoutFn = vi.fn((fn: () => void, ms: number) =>
-        setTimeout(fn, ms),
-      );
-
-      const cleanup = createBookmarksEffect({
-        bookmarkRetryCount: 2,
-        getBookmarksFn: async () => err('INTERNAL_ERROR', 'Boom'),
-        setBookmarkedQuestionIds: vi.fn(),
-        setBookmarkStatus: vi.fn(),
-        setBookmarkRetryCount: vi.fn(),
-        setTimeoutFn,
-        clearTimeoutFn: vi.fn(),
-        logError: vi.fn(),
-      });
-
-      await new Promise((resolve) => setTimeout(resolve, 0));
-      expect(setTimeoutFn).not.toHaveBeenCalled();
-
-      cleanup();
-    });
-
-    it('clears any scheduled timeout on cleanup', async () => {
-      vi.useFakeTimers();
-      try {
-        const setTimeoutFn = vi.fn((fn: () => void, ms: number) =>
-          setTimeout(fn, ms),
-        );
-        const clearTimeoutFn = vi.fn((id: ReturnType<typeof setTimeout>) =>
-          clearTimeout(id),
-        );
-
-        const cleanup = createBookmarksEffect({
-          bookmarkRetryCount: 0,
-          getBookmarksFn: async () => err('INTERNAL_ERROR', 'Boom'),
-          setBookmarkedQuestionIds: vi.fn(),
-          setBookmarkStatus: vi.fn(),
-          setBookmarkRetryCount: vi.fn(),
-          setTimeoutFn,
-          clearTimeoutFn,
-          logError: vi.fn(),
-        });
-
-        await vi.advanceTimersByTimeAsync(0);
-        cleanup();
-
-        expect(clearTimeoutFn).toHaveBeenCalledTimes(1);
-      } finally {
-        vi.useRealTimers();
-      }
-    });
-
-    it('returns no state updates when cleaned up before resolve', async () => {
-      const setBookmarkedQuestionIds = vi.fn();
-      const setBookmarkStatus = vi.fn();
-
-      let resolveBookmarks:
-        | ((
-            value: ActionResult<{ rows: Array<{ questionId: string }> }>,
-          ) => void)
-        | undefined;
-      const pending = new Promise<
-        ActionResult<{ rows: Array<{ questionId: string }> }>
-      >((res) => {
-        resolveBookmarks = res;
-      });
-      if (!resolveBookmarks) throw new Error('Expected resolve function');
-
-      const getBookmarksFn = vi.fn(async () => pending);
-
-      const cleanup = createBookmarksEffect({
-        bookmarkRetryCount: 0,
-        getBookmarksFn,
-        setBookmarkedQuestionIds,
-        setBookmarkStatus,
-        setBookmarkRetryCount: vi.fn(),
-        logError: vi.fn(),
-      });
-
-      cleanup();
-      resolveBookmarks(
-        ok({
-          rows: [{ questionId: 'q_1' }],
-        }),
-      );
-
-      await new Promise((resolve) => setTimeout(resolve, 0));
-
-      expect(setBookmarkedQuestionIds).not.toHaveBeenCalled();
-      expect(setBookmarkStatus).toHaveBeenCalledTimes(1);
-      expect(setBookmarkStatus).toHaveBeenCalledWith('loading');
-    });
-
-    it('sets error state when getBookmarksFn throws', async () => {
-      vi.useFakeTimers();
-      try {
-        const setBookmarkStatus = vi.fn();
-        const setTimeoutFn = vi.fn((fn: () => void, ms: number) =>
-          setTimeout(fn, ms),
-        );
-        const logError = vi.fn();
-
-        const cleanup = createBookmarksEffect({
-          bookmarkRetryCount: 0,
-          getBookmarksFn: async () => {
-            throw new Error('Boom');
-          },
-          setBookmarkedQuestionIds: vi.fn(),
-          setBookmarkStatus,
-          setBookmarkRetryCount: vi.fn(),
-          setTimeoutFn,
-          clearTimeoutFn: vi.fn(),
-          logError,
-        });
-
-        await vi.advanceTimersByTimeAsync(0);
-
-        expect(logError).toHaveBeenCalledWith(
-          'Failed to load bookmarks',
-          expect.any(Error),
-        );
-        expect(setBookmarkStatus).toHaveBeenCalledWith('error');
-        expect(setTimeoutFn).toHaveBeenCalledWith(expect.any(Function), 1000);
-
-        cleanup();
-      } finally {
-        vi.useRealTimers();
-      }
-    });
-
-    it('does not log expected business bookmark-load failures', async () => {
-      vi.useFakeTimers();
-      try {
-        const setBookmarkStatus = vi.fn();
-        const logError = vi.fn();
-
-        createBookmarksEffect({
-          bookmarkRetryCount: 0,
-          getBookmarksFn: async () =>
-            err('UNAUTHENTICATED', 'Authentication required'),
-          setBookmarkedQuestionIds: vi.fn(),
-          setBookmarkStatus,
-          setBookmarkRetryCount: vi.fn(),
-          setTimeoutFn: vi.fn((fn: () => void, ms: number) =>
-            setTimeout(fn, ms),
-          ),
-          clearTimeoutFn: vi.fn(),
-          logError,
-        });
-
-        await vi.advanceTimersByTimeAsync(0);
-
-        expect(setBookmarkStatus).toHaveBeenLastCalledWith('error');
-        expect(logError).not.toHaveBeenCalled();
-      } finally {
-        vi.useRealTimers();
-      }
-    });
-  });
-
-  describe('submitAnswerForQuestion', () => {
-    it('submits answer and sets result on success', async () => {
-      const submitAnswerFn = vi.fn(async () =>
-        ok({
-          attemptId: 'attempt_1',
-          isCorrect: true,
-          correctChoiceId: 'choice_1',
-          explanationMd: 'Because...',
-          referenceMd: null,
-          choiceExplanations: [],
-        } satisfies SubmitAnswerOutput),
-      );
-
-      const setLoadState = vi.fn();
-      const setSubmitResult = vi.fn();
-
-      await submitAnswerForQuestion({
-        question: createNextQuestion(),
-        selectedChoiceId: 'choice_1',
-        questionLoadedAtMs: 1000,
-        submitIdempotencyKey: 'idem_1',
-        submitAnswerFn,
-        nowMs: () => 6000,
-        setLoadState,
-        setSubmitResult,
-      });
-
-      expect(submitAnswerFn).toHaveBeenCalledWith({
-        questionId: 'q_1',
-        choiceId: 'choice_1',
-        idempotencyKey: 'idem_1',
-        timeSpentSeconds: 5,
-      });
-      expect(setSubmitResult).toHaveBeenCalledWith(
-        expect.objectContaining({ isCorrect: true }),
-        'q_1',
-      );
-      expect(setLoadState).not.toHaveBeenCalledWith({ status: 'loading' });
-    });
-
-    it('defaults timeSpentSeconds to 0 when loadedAt is null', async () => {
-      const submitAnswerFn = vi.fn(async () =>
-        ok({
-          attemptId: 'attempt_1',
-          isCorrect: true,
-          correctChoiceId: 'choice_1',
-          explanationMd: 'Because...',
-          referenceMd: null,
-          choiceExplanations: [],
-        } satisfies SubmitAnswerOutput),
-      );
-
-      await submitAnswerForQuestion({
-        question: createNextQuestion(),
-        selectedChoiceId: 'choice_1',
-        questionLoadedAtMs: null,
-        submitIdempotencyKey: 'idem_1',
-        submitAnswerFn,
-        nowMs: () => 0,
-        setLoadState: vi.fn(),
-        setSubmitResult: vi.fn(),
-      });
-
-      expect(submitAnswerFn).toHaveBeenCalledWith(
-        expect.objectContaining({ timeSpentSeconds: 0 }),
-      );
-    });
-
-    it('computes timeSpentSeconds when loadedAt is 0', async () => {
-      const submitAnswerFn = vi.fn(async () =>
-        ok({
-          attemptId: 'attempt_1',
-          isCorrect: true,
-          correctChoiceId: 'choice_1',
-          explanationMd: 'Because...',
-          referenceMd: null,
-          choiceExplanations: [],
-        } satisfies SubmitAnswerOutput),
-      );
-
-      await submitAnswerForQuestion({
-        question: createNextQuestion(),
-        selectedChoiceId: 'choice_1',
-        questionLoadedAtMs: 0,
-        submitIdempotencyKey: 'idem_1',
-        submitAnswerFn,
-        nowMs: () => 1500,
-        setLoadState: vi.fn(),
-        setSubmitResult: vi.fn(),
-      });
-
-      expect(submitAnswerFn).toHaveBeenCalledWith(
-        expect.objectContaining({ timeSpentSeconds: 1 }),
-      );
-    });
-
-    it('sets error state when submit fails', async () => {
-      const setLoadState = vi.fn();
-
-      await submitAnswerForQuestion({
-        question: createNextQuestion(),
-        selectedChoiceId: 'choice_1',
-        questionLoadedAtMs: 0,
-        submitIdempotencyKey: 'idem_1',
-        submitAnswerFn: async () => err('INTERNAL_ERROR', 'Boom'),
-        nowMs: () => 0,
-        setLoadState,
-        setSubmitResult: vi.fn(),
-      });
-
-      expect(setLoadState).toHaveBeenCalledWith({
-        status: 'error',
-        message: 'Boom',
-      });
-    });
-
-    it('sets error state when submit throws', async () => {
-      const setLoadState = vi.fn();
-
-      await submitAnswerForQuestion({
-        question: createNextQuestion(),
-        selectedChoiceId: 'choice_1',
-        questionLoadedAtMs: 0,
-        submitIdempotencyKey: 'idem_1',
-        submitAnswerFn: async () => {
-          throw new Error('Boom');
-        },
-        nowMs: () => 0,
-        setLoadState,
-        setSubmitResult: vi.fn(),
-      });
-
-      expect(setLoadState).toHaveBeenCalledWith({
-        status: 'error',
-        message: 'Boom',
-      });
-    });
-
-    it('returns no state updates when unmounted during submitAnswerForQuestion', async () => {
-      const deferred = createDeferred<ActionResult<SubmitAnswerOutput>>();
-      let mounted = true;
-
-      const setLoadState = vi.fn();
-      const setSubmitResult = vi.fn();
-
-      const promise = submitAnswerForQuestion({
-        question: createNextQuestion(),
-        selectedChoiceId: 'choice_1',
-        questionLoadedAtMs: 0,
-        submitIdempotencyKey: 'idem_1',
-        submitAnswerFn: async () => deferred.promise,
-        nowMs: () => 0,
-        setLoadState,
-        setSubmitResult,
-        isMounted: () => mounted,
-      });
-
-      mounted = false;
-      deferred.resolve(
-        ok({
-          attemptId: 'attempt_1',
-          isCorrect: true,
-          correctChoiceId: 'choice_1',
-          explanationMd: 'Because...',
-          referenceMd: null,
-          choiceExplanations: [],
-        } satisfies SubmitAnswerOutput),
-      );
-      await promise;
-
-      expect(setSubmitResult).not.toHaveBeenCalled();
-      expect(setLoadState).not.toHaveBeenCalledWith({ status: 'ready' });
-    });
-
-    it('returns no state updates when submit response is stale', async () => {
-      const deferred = createDeferred<ActionResult<SubmitAnswerOutput>>();
-      const setLoadState = vi.fn();
-      const setSubmitResult = vi.fn();
-
-      const promise = submitAnswerForQuestion({
-        question: createNextQuestion(),
-        selectedChoiceId: 'choice_1',
-        questionLoadedAtMs: 0,
-        submitIdempotencyKey: 'idem_1',
-        submitAnswerFn: async () => deferred.promise,
-        nowMs: () => 0,
-        setLoadState,
-        setSubmitResult,
-        createRequestSequenceId: () => 1,
-        isLatestRequest: () => false,
-      });
-
-      deferred.resolve(
-        ok({
-          attemptId: 'attempt_1',
-          isCorrect: true,
-          correctChoiceId: 'choice_1',
-          explanationMd: 'Because...',
-          referenceMd: null,
-          choiceExplanations: [],
-        } satisfies SubmitAnswerOutput),
-      );
-      await promise;
-
-      expect(setSubmitResult).not.toHaveBeenCalled();
-      expect(setLoadState).not.toHaveBeenCalled();
-    });
-  });
-
-  describe('toggleBookmarkForQuestion', () => {
-    it('toggles bookmark, forwards idempotency key, and rotates key on success', async () => {
-      let ids = new Set<string>(['other']);
-      const setBookmarkedQuestionIds = vi.fn(
-        (next: Set<string> | ((prev: Set<string>) => Set<string>)) => {
-          ids = typeof next === 'function' ? next(ids) : next;
-        },
-      );
-      const onBookmarkToggled = vi.fn();
-      const setBookmarkIdempotencyKey = vi.fn();
-      const toggleBookmarkFn = vi.fn(async () => ok({ bookmarked: true }));
-
-      await toggleBookmarkForQuestion({
-        question: createNextQuestion(),
-        bookmarkIdempotencyKey: 'idem_1',
-        createIdempotencyKey: () => 'idem_2',
-        setBookmarkIdempotencyKey,
-        toggleBookmarkFn,
-        setBookmarkStatus: vi.fn(),
-        setBookmarkedQuestionIds,
-        onBookmarkToggled,
-      });
-
-      expect(toggleBookmarkFn).toHaveBeenCalledWith({
-        questionId: 'q_1',
-        idempotencyKey: 'idem_1',
-      });
-      expect(ids.has('q_1')).toBe(true);
-      expect(onBookmarkToggled).toHaveBeenCalledWith(true);
-      expect(setBookmarkIdempotencyKey).toHaveBeenCalledWith('idem_2');
-    });
-
-    it('sets error state when toggle fails', async () => {
-      const setBookmarkStatus = vi.fn();
-      const onBookmarkToggled = vi.fn();
-
-      await toggleBookmarkForQuestion({
-        question: createNextQuestion(),
-        toggleBookmarkFn: async () => err('INTERNAL_ERROR', 'Boom'),
-        setBookmarkStatus,
-        setBookmarkedQuestionIds: vi.fn(),
-        onBookmarkToggled,
-      });
-
-      expect(setBookmarkStatus).toHaveBeenCalledWith('error');
-      expect(onBookmarkToggled).not.toHaveBeenCalled();
-    });
-
-    it('removes the question id when bookmark is removed', async () => {
-      let ids = new Set<string>(['q_1', 'other']);
-
-      const setBookmarkedQuestionIds = vi.fn(
-        (next: Set<string> | ((prev: Set<string>) => Set<string>)) => {
-          ids = typeof next === 'function' ? next(ids) : next;
-        },
-      );
-
-      const setBookmarkStatus = vi.fn();
-      const onBookmarkToggled = vi.fn();
-
-      await toggleBookmarkForQuestion({
-        question: createNextQuestion(),
-        toggleBookmarkFn: async () => ok({ bookmarked: false }),
-        setBookmarkStatus,
-        setBookmarkedQuestionIds,
-        onBookmarkToggled,
-      });
-
-      expect(ids.has('q_1')).toBe(false);
-      expect(onBookmarkToggled).toHaveBeenCalledWith(false);
-      expect(setBookmarkStatus).toHaveBeenLastCalledWith('idle');
-    });
-
-    it('sets error state when toggle throws', async () => {
-      const setBookmarkStatus = vi.fn();
-      const onBookmarkToggled = vi.fn();
-      const onBookmarkError = vi.fn();
-
-      await toggleBookmarkForQuestion({
-        question: createNextQuestion(),
-        toggleBookmarkFn: async () => {
-          throw new Error('Boom');
-        },
-        setBookmarkStatus,
-        setBookmarkedQuestionIds: vi.fn(),
-        onBookmarkToggled,
-        onBookmarkError,
-      });
-
-      expect(setBookmarkStatus).toHaveBeenLastCalledWith('error');
-      expect(onBookmarkToggled).not.toHaveBeenCalled();
-      expect(onBookmarkError).toHaveBeenCalledTimes(1);
-      expect(onBookmarkError).toHaveBeenCalledWith(
-        'Failed to save bookmark. Please try again.',
-      );
-    });
-
-    it('logs thrown toggle errors while preserving generic error UI state', async () => {
-      const error = new Error('Boom');
-      const logError = vi.fn();
-      const setBookmarkStatus = vi.fn();
-      const onBookmarkError = vi.fn();
-
-      await toggleBookmarkForQuestion({
-        question: createNextQuestion(),
-        toggleBookmarkFn: async () => {
-          throw error;
-        },
-        setBookmarkStatus,
-        setBookmarkedQuestionIds: vi.fn(),
-        onBookmarkError,
-        logError,
-      });
-
-      expect(logError).toHaveBeenCalledWith('Failed to toggle bookmark', error);
-      expect(setBookmarkStatus).toHaveBeenLastCalledWith('error');
-      expect(onBookmarkError).toHaveBeenCalledWith(
-        'Failed to save bookmark. Please try again.',
-      );
-    });
-
-    it('preserves generic bookmark error UI when logError throws for thrown errors', async () => {
-      const error = new Error('Boom');
-      const logError = vi.fn(() => {
-        throw new Error('logger failed');
-      });
-      const setBookmarkStatus = vi.fn();
-      const onBookmarkError = vi.fn();
-
-      await expect(
-        toggleBookmarkForQuestion({
-          question: createNextQuestion(),
-          toggleBookmarkFn: async () => {
-            throw error;
-          },
-          setBookmarkStatus,
-          setBookmarkedQuestionIds: vi.fn(),
-          onBookmarkError,
-          logError,
-        }),
-      ).resolves.toBeUndefined();
-
-      expect(logError).toHaveBeenCalledWith('Failed to toggle bookmark', error);
-      expect(setBookmarkStatus).toHaveBeenLastCalledWith('error');
-      expect(onBookmarkError).toHaveBeenCalledWith(
-        'Failed to save bookmark. Please try again.',
-      );
-    });
-
-    it('invokes error callback when toggle controller returns an error result', async () => {
-      const setBookmarkStatus = vi.fn();
-      const onBookmarkError = vi.fn();
-
-      await toggleBookmarkForQuestion({
-        question: createNextQuestion(),
-        toggleBookmarkFn: async () => err('INTERNAL_ERROR', 'Boom'),
-        setBookmarkStatus,
-        setBookmarkedQuestionIds: vi.fn(),
-        onBookmarkError,
-      });
-
-      expect(setBookmarkStatus).toHaveBeenLastCalledWith('error');
-      expect(onBookmarkError).toHaveBeenCalledTimes(1);
-      expect(onBookmarkError).toHaveBeenCalledWith(
-        'Failed to save bookmark. Please try again.',
-      );
-    });
-
-    it('logs structured toggle failures while preserving generic error UI state', async () => {
-      const structuredError = {
-        code: 'INTERNAL_ERROR',
-        message: 'Boom',
-      } as const;
-      const logError = vi.fn();
-      const setBookmarkStatus = vi.fn();
-      const onBookmarkError = vi.fn();
-
-      await toggleBookmarkForQuestion({
-        question: createNextQuestion(),
-        toggleBookmarkFn: async () => ({
-          ok: false,
-          error: structuredError,
-        }),
-        setBookmarkStatus,
-        setBookmarkedQuestionIds: vi.fn(),
-        onBookmarkError,
-        logError,
-      });
-
-      expect(logError).toHaveBeenCalledWith(
-        'Failed to toggle bookmark',
-        structuredError,
-      );
-      expect(setBookmarkStatus).toHaveBeenLastCalledWith('error');
-      expect(onBookmarkError).toHaveBeenCalledWith(
-        'Failed to save bookmark. Please try again.',
-      );
-    });
-
-    it('does not log expected business toggle failures while preserving generic error UI state', async () => {
-      const structuredError = {
-        code: 'UNAUTHENTICATED',
-        message: 'Authentication required',
-      } as const;
-      const logError = vi.fn();
-      const setBookmarkStatus = vi.fn();
-      const onBookmarkError = vi.fn();
-
-      await toggleBookmarkForQuestion({
-        question: createNextQuestion(),
-        toggleBookmarkFn: async () => ({
-          ok: false,
-          error: structuredError,
-        }),
-        setBookmarkStatus,
-        setBookmarkedQuestionIds: vi.fn(),
-        onBookmarkError,
-        logError,
-      });
-
-      expect(logError).not.toHaveBeenCalled();
-      expect(setBookmarkStatus).toHaveBeenLastCalledWith('error');
-      expect(onBookmarkError).toHaveBeenCalledWith(
-        'Failed to save bookmark. Please try again.',
-      );
-    });
-
-    it('preserves generic bookmark error UI when logError throws for structured failures', async () => {
-      const structuredError = {
-        code: 'INTERNAL_ERROR',
-        message: 'Boom',
-      } as const;
-      const logError = vi.fn(() => {
-        throw new Error('logger failed');
-      });
-      const setBookmarkStatus = vi.fn();
-      const onBookmarkError = vi.fn();
-
-      await expect(
-        toggleBookmarkForQuestion({
-          question: createNextQuestion(),
-          toggleBookmarkFn: async () => ({
-            ok: false,
-            error: structuredError,
-          }),
-          setBookmarkStatus,
-          setBookmarkedQuestionIds: vi.fn(),
-          onBookmarkError,
-          logError,
-        }),
-      ).resolves.toBeUndefined();
-
-      expect(logError).toHaveBeenCalledWith(
-        'Failed to toggle bookmark',
-        structuredError,
-      );
-      expect(setBookmarkStatus).toHaveBeenLastCalledWith('error');
-      expect(onBookmarkError).toHaveBeenCalledWith(
-        'Failed to save bookmark. Please try again.',
-      );
-    });
-
-    it('returns no state updates when unmounted during toggleBookmarkForQuestion', async () => {
-      const deferred = createDeferred<ActionResult<{ bookmarked: boolean }>>();
-      let mounted = true;
-
-      const setBookmarkStatus = vi.fn();
-      const setBookmarkedQuestionIds = vi.fn();
-      const onBookmarkToggled = vi.fn();
-
-      const promise = toggleBookmarkForQuestion({
-        question: createNextQuestion(),
-        toggleBookmarkFn: async () => deferred.promise,
-        setBookmarkStatus,
-        setBookmarkedQuestionIds,
-        onBookmarkToggled,
-        isMounted: () => mounted,
-      });
-
-      mounted = false;
-      deferred.resolve(ok({ bookmarked: true }));
-      await promise;
-
-      expect(setBookmarkedQuestionIds).not.toHaveBeenCalled();
-      expect(onBookmarkToggled).not.toHaveBeenCalled();
-      expect(setBookmarkStatus).not.toHaveBeenCalledWith('idle');
-    });
-
-    it('logs thrown toggle errors after unmount without applying error UI state', async () => {
-      const deferred = createDeferred<ActionResult<{ bookmarked: boolean }>>();
-      const error = new Error('Boom');
-      let mounted = true;
-
-      const logError = vi.fn();
-      const setBookmarkStatus = vi.fn();
-      const onBookmarkError = vi.fn();
-
-      const promise = toggleBookmarkForQuestion({
-        question: createNextQuestion(),
-        toggleBookmarkFn: async () => deferred.promise,
-        setBookmarkStatus,
-        setBookmarkedQuestionIds: vi.fn(),
-        onBookmarkError,
-        logError,
-        isMounted: () => mounted,
-      });
-
-      mounted = false;
-      deferred.reject(error);
-      await promise;
-
-      expect(logError).toHaveBeenCalledWith('Failed to toggle bookmark', error);
-      expect(onBookmarkError).not.toHaveBeenCalled();
-      expect(setBookmarkStatus).not.toHaveBeenCalledWith('error');
-    });
-
-    it('logs structured toggle failures after unmount without applying error UI state', async () => {
-      const deferred = createDeferred<ActionResult<{ bookmarked: boolean }>>();
-      const structuredError = {
-        code: 'INTERNAL_ERROR',
-        message: 'Boom',
-      } as const;
-      let mounted = true;
-
-      const logError = vi.fn();
-      const setBookmarkStatus = vi.fn();
-      const onBookmarkError = vi.fn();
-
-      const promise = toggleBookmarkForQuestion({
-        question: createNextQuestion(),
-        toggleBookmarkFn: async () => deferred.promise,
-        setBookmarkStatus,
-        setBookmarkedQuestionIds: vi.fn(),
-        onBookmarkError,
-        logError,
-        isMounted: () => mounted,
-      });
-
-      mounted = false;
-      deferred.resolve({
-        ok: false,
-        error: structuredError,
-      });
-      await promise;
-
-      expect(logError).toHaveBeenCalledWith(
-        'Failed to toggle bookmark',
-        structuredError,
-      );
-      expect(onBookmarkError).not.toHaveBeenCalled();
-      expect(setBookmarkStatus).not.toHaveBeenCalledWith('error');
-    });
-  });
-
-  describe('selectChoiceIfAllowed', () => {
-    it('does nothing when submitResult exists', () => {
-      const setSelectedChoiceId = vi.fn();
-
-      const changed = selectChoiceIfAllowed(
-        {
-          isAnswered: false,
-          submitResult: {
-            attemptId: 'attempt_1',
-            isCorrect: true,
-            correctChoiceId: 'choice_1',
-            explanationMd: 'Because...',
-            referenceMd: null,
-            choiceExplanations: [],
-          },
-        },
-        setSelectedChoiceId,
-        'choice_1',
-      );
-
-      expect(setSelectedChoiceId).not.toHaveBeenCalled();
-      expect(changed).toBe(false);
-    });
-
-    it('sets the choice when no submitResult exists', () => {
-      const setSelectedChoiceId = vi.fn();
-
-      const changed = selectChoiceIfAllowed(
-        { isAnswered: false, submitResult: null },
-        setSelectedChoiceId,
-        'choice_1',
-      );
-
-      expect(setSelectedChoiceId).toHaveBeenCalledWith('choice_1');
-      expect(changed).toBe(true);
-    });
-
-    it('does nothing when question is already answered', () => {
-      const setSelectedChoiceId = vi.fn();
-
-      const changed = selectChoiceIfAllowed(
-        { isAnswered: true, submitResult: null },
-        setSelectedChoiceId,
-        'choice_1',
-      );
-
-      expect(setSelectedChoiceId).not.toHaveBeenCalled();
-      expect(changed).toBe(false);
-    });
-  });
-
-  describe('handleSessionModeChange', () => {
-    it('sets mode only for allowed values', () => {
-      const setSessionMode = vi.fn();
-
-      handleSessionModeChange(setSessionMode, { target: { value: 'tutor' } });
-      handleSessionModeChange(setSessionMode, { target: { value: 'exam' } });
-      handleSessionModeChange(setSessionMode, { target: { value: 'nope' } });
-
-      expect(setSessionMode).toHaveBeenCalledTimes(2);
-      expect(setSessionMode).toHaveBeenNthCalledWith(1, 'tutor');
-      expect(setSessionMode).toHaveBeenNthCalledWith(2, 'exam');
-    });
-  });
-
-  describe('handleSessionCountChange', () => {
-    it('stores the raw input value and sets the count from numeric input', () => {
-      const setSessionCountInputValue = vi.fn();
-      const setSessionCount = vi.fn();
-
-      handleSessionCountChange(setSessionCountInputValue, setSessionCount, {
-        target: { value: '12' },
-      });
-
-      expect(setSessionCountInputValue).toHaveBeenCalledWith('12');
-      expect(setSessionCount).toHaveBeenCalledWith(12);
-    });
-
-    it('stores the raw input value and clamps finite values within range', () => {
-      const setSessionCountInputValue = vi.fn();
-      const setSessionCount = vi.fn();
-
-      handleSessionCountChange(setSessionCountInputValue, setSessionCount, {
-        target: { value: '0' },
-      });
-      handleSessionCountChange(setSessionCountInputValue, setSessionCount, {
-        target: { value: '-1' },
-      });
-
-      expect(setSessionCountInputValue).toHaveBeenNthCalledWith(1, '0');
-      expect(setSessionCountInputValue).toHaveBeenNthCalledWith(2, '-1');
-      expect(setSessionCount).toHaveBeenNthCalledWith(1, SESSION_COUNT_MIN);
-      expect(setSessionCount).toHaveBeenNthCalledWith(2, SESSION_COUNT_MIN);
-    });
-
-    it('stores the raw input value and clamps to maximum when value is above range', () => {
-      const setSessionCountInputValue = vi.fn();
-      const setSessionCount = vi.fn();
-
-      handleSessionCountChange(setSessionCountInputValue, setSessionCount, {
-        target: { value: '101' },
-      });
-
-      expect(setSessionCountInputValue).toHaveBeenCalledWith('101');
-      expect(setSessionCount).toHaveBeenCalledWith(SESSION_COUNT_MAX);
-    });
-
-    it('stores raw input without changing the count when the field is cleared or invalid', () => {
-      const setSessionCountInputValue = vi.fn();
-      const setSessionCount = vi.fn();
-
-      handleSessionCountChange(setSessionCountInputValue, setSessionCount, {
-        target: { value: '' },
-      });
-      handleSessionCountChange(setSessionCountInputValue, setSessionCount, {
-        target: { value: 'NaN' },
-      });
-
-      expect(setSessionCountInputValue).toHaveBeenNthCalledWith(1, '');
-      expect(setSessionCountInputValue).toHaveBeenNthCalledWith(2, 'NaN');
-      expect(setSessionCount).not.toHaveBeenCalled();
-    });
-  });
-
-  describe('handleSessionCountBlur', () => {
-    it('resets the raw input string to the canonical numeric count on blur', () => {
-      const setSessionCountInputValue = vi.fn();
-
-      handleSessionCountBlur(20, setSessionCountInputValue);
-
-      expect(setSessionCountInputValue).toHaveBeenCalledWith('20');
-    });
-  });
-
-  describe('startSession', () => {
-    it('sets error state when controller fails', async () => {
-      const setSessionStartStatus = vi.fn();
-      const setSessionStartError = vi.fn();
-      const setIdempotencyKey = vi.fn();
-
-      await startSession({
-        sessionMode: 'tutor',
-        sessionCount: 20,
-        filters: {
-          tagSlugs: ['alcohol'],
-          difficulty: null,
-          status: 'unanswered',
-        },
-        idempotencyKey: 'idem_1',
-        createIdempotencyKey: () => 'idem_2',
-        setIdempotencyKey,
-        startPracticeSessionFn: async () => err('NOT_FOUND', 'No questions'),
-        setSessionStartStatus,
-        setSessionStartError,
-        navigateTo: vi.fn(),
-      });
-
-      expect(setSessionStartStatus).toHaveBeenCalledWith('error');
-      expect(setSessionStartError).toHaveBeenCalledWith('No questions');
-      expect(setIdempotencyKey).toHaveBeenCalledWith('idem_2');
-    });
-
-    it('navigates to the session route on success without toast params when counts match', async () => {
-      const startPracticeSessionFn = vi.fn(async () =>
-        ok({ sessionId: 'session-1', requestedCount: 10, actualCount: 10 }),
-      );
-      const navigateTo = vi.fn();
-      const setIdempotencyKey = vi.fn();
-
-      await startSession({
-        sessionMode: 'exam',
-        sessionCount: 10,
-        filters: {
-          tagSlugs: ['opioids'],
-          difficulty: 'hard',
-          status: 'unanswered',
-        },
-        idempotencyKey: 'idem_1',
-        createIdempotencyKey: () => 'idem_2',
-        setIdempotencyKey,
-        startPracticeSessionFn,
-        setSessionStartStatus: vi.fn(),
-        setSessionStartError: vi.fn(),
-        navigateTo,
-      });
-
-      expect(startPracticeSessionFn).toHaveBeenCalledWith({
-        mode: 'exam',
-        count: 10,
-        idempotencyKey: 'idem_1',
-        tagSlugs: ['opioids'],
-        difficulties: ['hard'],
-        statuses: ['unanswered'],
-      });
-      expect(navigateTo).toHaveBeenCalledWith(
-        toPracticeSessionRoute('session-1'),
-      );
-      expect(setIdempotencyKey).not.toHaveBeenCalled();
-    });
-
-    it('includes requested/actual counts in the session_started toast when fewer questions are available than requested', async () => {
-      const startPracticeSessionFn = vi.fn(async () =>
-        ok({ sessionId: 'session-1', requestedCount: 50, actualCount: 30 }),
-      );
-      const navigateTo = vi.fn();
-
-      await startSession({
-        sessionMode: 'exam',
-        sessionCount: 50,
-        filters: { tagSlugs: [], difficulty: null, status: 'unanswered' },
-        idempotencyKey: 'idem_1',
-        createIdempotencyKey: () => 'idem_2',
-        setIdempotencyKey: vi.fn(),
-        startPracticeSessionFn,
-        setSessionStartStatus: vi.fn(),
-        setSessionStartError: vi.fn(),
-        navigateTo,
-      });
-
-      expect(navigateTo).toHaveBeenCalledWith(
-        `${toPracticeSessionRoute('session-1')}?toast=session_started&requestedCount=50&actualCount=30`,
-      );
-    });
-
-    it('sets error state when controller throws', async () => {
-      const setSessionStartStatus = vi.fn();
-      const setSessionStartError = vi.fn();
-      const setIdempotencyKey = vi.fn();
-
-      await startSession({
-        sessionMode: 'tutor',
-        sessionCount: 20,
-        filters: {
-          tagSlugs: ['alcohol'],
-          difficulty: null,
-          status: 'unanswered',
-        },
-        idempotencyKey: 'idem_1',
-        createIdempotencyKey: () => 'idem_2',
-        setIdempotencyKey,
-        startPracticeSessionFn: async () => {
-          throw new Error('Boom');
-        },
-        setSessionStartStatus,
-        setSessionStartError,
-        navigateTo: vi.fn(),
-      });
-
-      expect(setSessionStartStatus).toHaveBeenCalledWith('error');
-      expect(setSessionStartError).toHaveBeenCalledWith('Boom');
-      expect(setIdempotencyKey).toHaveBeenCalledWith('idem_2');
-    });
-
-    it('reports thrown session start errors while preserving error UI state', async () => {
-      const error = new Error('Boom');
-      const reportError = vi.fn();
-      const setSessionStartStatus = vi.fn();
-      const setSessionStartError = vi.fn();
-      const setIdempotencyKey = vi.fn();
-
-      await startSession({
-        sessionMode: 'tutor',
-        sessionCount: 20,
-        filters: {
-          tagSlugs: ['alcohol'],
-          difficulty: null,
-          status: 'unanswered',
-        },
-        idempotencyKey: 'idem_1',
-        createIdempotencyKey: () => 'idem_2',
-        setIdempotencyKey,
-        startPracticeSessionFn: async () => {
-          throw error;
-        },
-        reportError,
-        setSessionStartStatus,
-        setSessionStartError,
-        navigateTo: vi.fn(),
-      });
-
-      expect(reportError).toHaveBeenCalledWith(error, {
-        action: 'startSession',
-      });
-      expect(setSessionStartStatus).toHaveBeenCalledWith('error');
-      expect(setSessionStartError).toHaveBeenCalledWith('Boom');
-      expect(setIdempotencyKey).toHaveBeenCalledWith('idem_2');
-    });
-
-    it('preserves session start error handling when reportError throws', async () => {
-      const error = new Error('Boom');
-      const reportError = vi.fn(() => {
-        throw new Error('reporter failed');
-      });
-      const setSessionStartStatus = vi.fn();
-      const setSessionStartError = vi.fn();
-      const setIdempotencyKey = vi.fn();
-
-      await expect(
-        startSession({
-          sessionMode: 'tutor',
-          sessionCount: 20,
-          filters: {
-            tagSlugs: ['alcohol'],
-            difficulty: null,
-            status: 'unanswered',
-          },
-          idempotencyKey: 'idem_1',
-          createIdempotencyKey: () => 'idem_2',
-          setIdempotencyKey,
-          startPracticeSessionFn: async () => {
-            throw error;
-          },
-          reportError,
-          setSessionStartStatus,
-          setSessionStartError,
-          navigateTo: vi.fn(),
-        }),
-      ).resolves.toBeUndefined();
-
-      expect(reportError).toHaveBeenCalledWith(error, {
-        action: 'startSession',
-      });
-      expect(setSessionStartStatus).toHaveBeenCalledWith('error');
-      expect(setSessionStartError).toHaveBeenCalledWith('Boom');
-      expect(setIdempotencyKey).toHaveBeenCalledWith('idem_2');
-    });
-
-    it('returns without navigating when unmounted during startSession', async () => {
-      const deferred =
-        createDeferred<
-          ActionResult<{
-            sessionId: string;
-            requestedCount: number;
-            actualCount: number;
-          }>
-        >();
-      let mounted = true;
-
-      const navigateTo = vi.fn();
-
-      const promise = startSession({
-        sessionMode: 'exam',
-        sessionCount: 10,
-        filters: {
-          tagSlugs: ['opioids'],
-          difficulty: 'hard',
-          status: 'unanswered',
-        },
-        idempotencyKey: 'idem_1',
-        createIdempotencyKey: () => 'idem_2',
-        setIdempotencyKey: vi.fn(),
-        startPracticeSessionFn: async () => deferred.promise,
-        setSessionStartStatus: vi.fn(),
-        setSessionStartError: vi.fn(),
-        navigateTo,
-        isMounted: () => mounted,
-      });
-
-      mounted = false;
-      deferred.resolve(
-        ok({ sessionId: 'session-1', requestedCount: 10, actualCount: 10 }),
-      );
-      await promise;
-
-      expect(navigateTo).not.toHaveBeenCalled();
-    });
-
-    it('reports thrown session start errors after unmount without applying error UI state', async () => {
-      const deferred =
-        createDeferred<
-          ActionResult<{
-            sessionId: string;
-            requestedCount: number;
-            actualCount: number;
-          }>
-        >();
-      const error = new Error('Boom');
-      let mounted = true;
-
-      const reportError = vi.fn();
-      const setSessionStartStatus = vi.fn();
-      const setSessionStartError = vi.fn();
-      const setIdempotencyKey = vi.fn();
-
-      const promise = startSession({
-        sessionMode: 'exam',
-        sessionCount: 10,
-        filters: {
-          tagSlugs: ['opioids'],
-          difficulty: 'hard',
-          status: 'unanswered',
-        },
-        idempotencyKey: 'idem_1',
-        createIdempotencyKey: () => 'idem_2',
-        setIdempotencyKey,
-        startPracticeSessionFn: async () => deferred.promise,
-        reportError,
-        setSessionStartStatus,
-        setSessionStartError,
-        navigateTo: vi.fn(),
-        isMounted: () => mounted,
-      });
-
-      mounted = false;
-      deferred.reject(error);
-      await promise;
-
-      expect(reportError).toHaveBeenCalledWith(error, {
-        action: 'startSession',
-      });
-      expect(setSessionStartStatus).not.toHaveBeenCalledWith('error');
-      expect(setSessionStartError).not.toHaveBeenCalledWith('Boom');
-      expect(setIdempotencyKey).not.toHaveBeenCalledWith('idem_2');
-    });
-  });
-});
-
-describe('practice-page-session-start handlers', () => {
-  it('rotates idempotency key when session mode changes', () => {
-    const setSessionMode = vi.fn();
-    const setIdempotencyKey = vi.fn();
-    const handler = createSessionModeChangeHandler({
-      setSessionMode,
-      setIdempotencyKey,
-      createIdempotencyKey: () => 'idem_2',
-    });
-
-    handler('exam');
-
-    expect(setSessionMode).toHaveBeenCalledWith('exam');
-    expect(setIdempotencyKey).toHaveBeenCalledWith('idem_2');
-  });
-
-  it('ignores invalid session mode values', () => {
-    const setSessionMode = vi.fn();
-    const setIdempotencyKey = vi.fn();
-    const handler = createSessionModeChangeHandler({
-      setSessionMode,
-      setIdempotencyKey,
-      createIdempotencyKey: () => 'idem_2',
-    });
-
-    handler('nope');
-
-    expect(setSessionMode).not.toHaveBeenCalled();
-    expect(setIdempotencyKey).not.toHaveBeenCalled();
-  });
-
-  it('rotates idempotency key when session count changes', () => {
-    const setSessionCount = vi.fn();
-    const setSessionCountInputValue = vi.fn();
-    const setIdempotencyKey = vi.fn();
-    const handler = createSessionCountChangeHandler({
-      setSessionCountInputValue,
-      setSessionCount,
-      setIdempotencyKey,
-      createIdempotencyKey: () => 'idem_2',
-    });
-
-    handler({ target: { value: '12' } });
-
-    expect(setSessionCountInputValue).toHaveBeenCalledWith('12');
-    expect(setSessionCount).toHaveBeenCalledWith(12);
-    expect(setIdempotencyKey).toHaveBeenCalledWith('idem_2');
-  });
-
-  it('does not rotate idempotency key when the session count field is cleared', () => {
-    const setSessionCount = vi.fn();
-    const setSessionCountInputValue = vi.fn();
-    const setIdempotencyKey = vi.fn();
-    const handler = createSessionCountChangeHandler({
-      setSessionCountInputValue,
-      setSessionCount,
-      setIdempotencyKey,
-      createIdempotencyKey: () => 'idem_2',
-    });
-
-    handler({ target: { value: '' } });
-
-    expect(setSessionCountInputValue).toHaveBeenCalledWith('');
-    expect(setSessionCount).not.toHaveBeenCalled();
-    expect(setIdempotencyKey).not.toHaveBeenCalled();
-  });
-
-  it('resets the raw input string to the canonical numeric count on blur', () => {
-    const setSessionCountInputValue = vi.fn();
-    const handler = createSessionCountBlurHandler({
-      sessionCount: 20,
-      setSessionCountInputValue,
-    });
-
-    handler();
-
-    expect(setSessionCountInputValue).toHaveBeenCalledWith('20');
-  });
-
-  it('applies tag toggles and rotates idempotency key', () => {
-    const setIdempotencyKey = vi.fn();
-    const setFilters = vi.fn();
-    const handler = createToggleTagHandler({
-      setFilters,
-      setIdempotencyKey,
-      createIdempotencyKey: () => 'idem_2',
-    });
-
-    handler('opioids');
-
-    expect(setIdempotencyKey).toHaveBeenCalledWith('idem_2');
-    const update = setFilters.mock.calls[0]?.[0];
-    if (typeof update !== 'function') {
-      throw new Error('Expected setFilters to receive an updater function');
-    }
-
-    expect(
-      update({ tagSlugs: [], difficulty: null, status: 'unanswered' }),
-    ).toEqual({
-      tagSlugs: ['opioids'],
-      difficulty: null,
-      status: 'unanswered',
-    });
-
-    expect(
-      update({
-        tagSlugs: ['opioids'],
-        difficulty: 'easy',
-        status: 'incorrect',
-      }),
-    ).toEqual({
-      tagSlugs: [],
-      difficulty: 'easy',
-      status: 'incorrect',
-    });
-  });
-
-  it('applies difficulty changes and rotates idempotency key', () => {
-    const setIdempotencyKey = vi.fn();
-    const setFilters = vi.fn();
-    const handler = createDifficultyChangeHandler({
-      setFilters,
-      setIdempotencyKey,
-      createIdempotencyKey: () => 'idem_2',
-    });
-
-    handler('hard');
-
-    expect(setIdempotencyKey).toHaveBeenCalledWith('idem_2');
-    const update = setFilters.mock.calls[0]?.[0];
-    if (typeof update !== 'function') {
-      throw new Error('Expected setFilters to receive an updater function');
-    }
-
-    expect(
-      update({ tagSlugs: [], difficulty: null, status: 'unanswered' }),
-    ).toEqual({
-      tagSlugs: [],
-      difficulty: 'hard',
-      status: 'unanswered',
-    });
-
-    expect(
-      update({ tagSlugs: [], difficulty: 'hard', status: 'unanswered' }),
-    ).toEqual({
-      tagSlugs: [],
-      difficulty: 'hard',
-      status: 'unanswered',
-    });
-  });
-
-  it('applies status changes and rotates idempotency key', () => {
-    const setIdempotencyKey = vi.fn();
-    const setFilters = vi.fn();
-    const handler = createStatusChangeHandler({
-      setFilters,
-      setIdempotencyKey,
-      createIdempotencyKey: () => 'idem_2',
-    });
-
-    handler('incorrect');
-
-    expect(setIdempotencyKey).toHaveBeenCalledWith('idem_2');
-    const update = setFilters.mock.calls[0]?.[0];
-    if (typeof update !== 'function') {
-      throw new Error('Expected setFilters to receive an updater function');
-    }
-
-    expect(
-      update({ tagSlugs: [], difficulty: null, status: 'unanswered' }),
-    ).toEqual({
-      tagSlugs: [],
-      difficulty: null,
-      status: 'incorrect',
-    });
-
-    expect(
-      update({ tagSlugs: [], difficulty: null, status: 'incorrect' }),
-    ).toEqual({
-      tagSlugs: [],
-      difficulty: null,
-      status: 'incorrect',
-    });
-  });
-});
